pkg/source: don't clone, just open existing
In the end it is almost certain that we will get -packages git source
from disk anyway and just manipulate it in ramdisk within nbuild because
cloning is computationally expensive.

Hence the clone should be externalized as we are not interested in a
usecase where the clone is done within nbuild.

Along with this drop the Url field because we are not using that
anymore.